fix(sync): address PR #1 round-1 review comments
- cli: add clap value parsers so bpm/sr/ppq/pulses must be positive
  and jitter-us non-negative; previously these flowed straight into
  pulse_train()'s assert! and panicked the CLI on bad input.
- detect: tighten hold_samples docstring to spell out the inclusive
  range that's blocked. Also widen start_index arithmetic to i128 so
  stream-global indices above i64::MAX no longer wrap to negatives.
- pll: floor the integrator clamp's lower bound just above -1.0 so
  default PllSettings (clamp_hz / nominal_freq_hz = 50/48 > 1) can no
  longer drive smoothed BPM negative. Add Pll::last_pulse_sample()
  getter and a regression test that walks the integrator hard against
  the floor.
- source: External::phase_at_sample(n) now projects analytically from
  the PLL's last observed pulse instead of ignoring n and returning
  the last-step snapshot. feed_samples leaves silent blocks alone --
  analytic projection handles the gap. New tests cover halfway-
  between-pulses, exact-cycle wrap, and the no-pulse-yet zero case.

+++ crates/core/src/sync/detect.rs
@@ -18,8 +18,10 @@ pub struct Peak {
 pub struct DetectorConfig {
     /// Minimum amplitude for a candidate to be considered a peak.
     pub threshold: f32,
-    /// Minimum number of samples between consecutive emitted peaks.
-    /// A peak at sample `n` blocks emission at samples `n+1 .. n+hold`.
+    /// Minimum number of samples between consecutive emitted peaks: a
+    /// peak at sample `n` suppresses any candidate at samples `n+1`
+    /// through `n + hold_samples - 1` inclusive; the next emission is
+    /// allowed at `n + hold_samples` or later.
     pub hold_samples: u32,
 }
 
@@ -103,8 +105,11 @@ impl PeakDetector {
                         (0.5 * (y_m1 - y_p1) as f64 / denom).clamp(-0.5, 0.5)
                     };
 
+                    // i128 keeps the arithmetic exact for any
+                    // u64 stream index — `start_index as i64` would
+                    // wrap above `i64::MAX`.
                     let centre_global =
-                        (start_index as i64) + (offset as i64) - 1;
+                        (start_index as i128) + (offset as i128) - 1;
                     let sample_index = centre_global as f64 + frac;
 
                     // Parabolic-fit apex amplitude.
